# Conflicts:
#	ROOT/ROOT.iml
#	mercury/src/main/db/athenadb.changelog-1.3.xml
#	mercury/src/main/db/mercurydb.changelog-1.2.xml
#	mercury/src/test/java/org/broadinstitute/gpinformatics/mercury/test/BaseEventTest.java

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/infrastructure/quote/QuoteServiceStub.java
@@ -51,14 +51,14 @@ public class QuoteServiceStub implements QuoteService {
     @Override
     public String registerNewWork(Quote quote, QuotePriceItem quotePriceItem, QuotePriceItem itemIsReplacing,
                                   Date reportedCompletionDate,
-                                  double numWorkUnits, String callbackUrl, String callbackParameterName,
+                                  BigDecimal numWorkUnits, String callbackUrl, String callbackParameterName,
                                   String callbackParameterValue, BigDecimal priceAdjustment) {
         return Integer.toString(workItemId++);
     }
 
     @Override
     public String registerNewSAPWork(Quote quote, QuotePriceItem quotePriceItem, QuotePriceItem itemIsReplacing,
-                                     Date reportedCompletionDate, double numWorkUnits, String callbackUrl,
+                                     Date reportedCompletionDate, BigDecimal numWorkUnits, String callbackUrl,
                                      String callbackParameterName, String callbackParameterValue,
                                      BigDecimal priceAdjustment) {
         return Integer.toString(workItemId++);
